+++ CHANGELOG.md
@@ -19,3 +19,6 @@
   - Support for setting data using Jinja2 templates.
   - **CHANGE OF BEHAVIOUR** `session.connect()` should not be `session.connect('yang-module-name')`. `session.get_root('yang-mdoule-name')` should be `session.get_root()`. This forces a single session per yang module.
   - FIX: stubdal to create /path[key='val']/key = val to match sysrepo implementation.
+  - **CHANGE OF BEHAVIOUR** `session.get_root()` renamed to `session.get_node()`
+  - **CHANGE OF BEHAVIOUR** `session.get_node()` does not take in `yang_location` this has now moved to `session.connect`
+